changed to ints

+++ lcm-types/cpp/spi_command_t.hpp
@@ -14,7 +14,7 @@
 class spi_command_t
 {
     public:
-        float      q_des_abad[4] = {1, 2, 3, 4};
+        int      q_des_abad[4] = {1, 2, 3, 4};
 
         // float      q_des_hip[4] = {5, 6, 7, 8};
